requirements.txt
django>=6.0,<7.0
djangorestframework>=3.15,<4.0
requests>=2.32.0
pillow>=12.0.0
qrcode>=8.0
gunicorn>=23.0.0
